All code was updated & cmd working without the path
