+++ free_array.c
@@ -13,5 +13,6 @@ void free_array(char **tokens)
 		free(tokens[i]);
 		i++;
 	}
+
 	free(tokens);
 }
